Added back confimation for dotfile overwrite

+++ .utils/symlinks.sh
@@ -20,12 +20,16 @@ main() {
 
         sourceFile="$(pwd)/$i"
         targetFile="$HOME/$(printf "%s" "$i" | sed "s/.*\/\(.*\)/\1/g")"
-
         if [ -e "$targetFile" ]; then
             if [ "$(readlink "$targetFile")" != "$sourceFile" ]; then
 
-                rm -rf "$targetFile"
-                execute "ln -fs $sourceFile $targetFile" "$targetFile → $sourceFile"
+                ask_for_confirmation "'$targetFile' already exists, do you want to overwrite it?"
+                if answer_is_yes; then
+                    rm -rf "$targetFile"
+                    execute "ln -fs $sourceFile $targetFile" "$targetFile → $sourceFile"
+                else
+                    print_error "$targetFile → $sourceFile"
+                fi
 
             else
                 print_success "$targetFile → $sourceFile"
@@ -33,6 +37,7 @@ main() {
         else
             execute "ln -fs $sourceFile $targetFile" "$targetFile → $sourceFile"
         fi
+
     done
 }
